commit selasa 19-04-2022
-Struktur Project HikariCP dapat digabung dengan rental mobil
-REST controller Biaya dan controller biaya dapat digabung
-Halaman webpage Biaya dapat tampil dengan REST Controller menggunakan Model and View

+++ src/main/java/com/rentalmobil/biaya/BiayaRepository.java
@@ -3,5 +3,5 @@ package com.rentalmobil.biaya;
 import org.springframework.data.repository.CrudRepository;
 
 public interface BiayaRepository extends CrudRepository<Biaya, Integer> {
-    public Long countById(Integer id);
+    public Integer countById(Integer id);
 }
